+++ packages/repo-tools/src/commands/api-reports/api-extractor.ts
@@ -65,9 +65,11 @@ import {
 } from '@microsoft/api-documenter/lib/markdown/CustomMarkdownEmitter';
 import { IMarkdownEmitterContext } from '@microsoft/api-documenter/lib/markdown/MarkdownEmitter';
 import { AstDeclaration } from '@microsoft/api-extractor/lib/analyzer/AstDeclaration';
+import { paths as cliPaths } from '../../lib/paths';
 
-const tmpDir = resolvePath(
-  process.cwd(),
+import minimatch from 'minimatch';
+
+const tmpDir = cliPaths.resolveTargetRoot(
   './node_modules/.cache/api-extractor',
 );
 
@@ -219,77 +221,8 @@ ApiReportGenerator.generateReviewFileContent =
     });
   };
 
-const PACKAGE_ROOTS = ['packages', 'plugins'];
-
-const ALLOW_WARNINGS = [
-  'packages/core-components',
-  'plugins/catalog',
-  'plugins/catalog-import',
-  'plugins/git-release-manager',
-  'plugins/jenkins',
-  'plugins/kubernetes',
-];
-
-async function resolvePackagePath(
-  packagePath: string,
-): Promise<string | undefined> {
-  const projectRoot = resolvePath(process.cwd());
-  const fullPackageDir = resolvePath(projectRoot, packagePath);
-
-  const stat = await fs.stat(fullPackageDir);
-  if (!stat.isDirectory()) {
-    return undefined;
-  }
-
-  try {
-    const packageJsonPath = join(fullPackageDir, 'package.json');
-    await fs.access(packageJsonPath);
-  } catch (_) {
-    return undefined;
-  }
-
-  return relativePath(projectRoot, fullPackageDir);
-}
-
-export async function findSpecificPackageDirs(unresolvedPackageDirs: string[]) {
-  const packageDirs = new Array<string>();
-
-  for (const unresolvedPackageDir of unresolvedPackageDirs) {
-    const packageDir = await resolvePackagePath(unresolvedPackageDir);
-    if (!packageDir) {
-      throw new Error(`'${unresolvedPackageDir}' is not a valid package path`);
-    }
-    packageDirs.push(packageDir);
-  }
-
-  if (packageDirs.length === 0) {
-    return undefined;
-  }
-
-  return packageDirs;
-}
-
-export async function findPackageDirs() {
-  const packageDirs = new Array<string>();
-  const projectRoot = resolvePath(process.cwd());
-
-  for (const packageRoot of PACKAGE_ROOTS) {
-    const dirs = await fs.readdir(resolvePath(projectRoot, packageRoot));
-    for (const dir of dirs) {
-      const packageDir = await resolvePackagePath(join(packageRoot, dir));
-      if (!packageDir) {
-        continue;
-      }
-
-      packageDirs.push(packageDir);
-    }
-  }
-
-  return packageDirs;
-}
-
 export async function createTemporaryTsConfig(includedPackageDirs: string[]) {
-  const path = resolvePath(process.cwd(), 'tsconfig.tmp.json');
+  const path = cliPaths.resolveTargetRoot('tsconfig.tmp.json');
 
   process.once('exit', () => {
     fs.removeSync(path);
@@ -375,6 +308,8 @@ interface ApiExtractionOptions {
   outputDir: string;
   isLocalBuild: boolean;
   tsconfigFilePath: string;
+  allowWarnings?: boolean | string[];
+  omitMessages?: string[];
 }
 
 export async function runApiExtraction({
@@ -382,25 +317,37 @@ export async function runApiExtraction({
   outputDir,
   isLocalBuild,
   tsconfigFilePath,
+  allowWarnings = false,
+  omitMessages = [],
 }: ApiExtractionOptions) {
   await fs.remove(outputDir);
 
   const entryPoints = packageDirs.map(packageDir => {
-    return resolvePath(
-      process.cwd(),
+    return cliPaths.resolveTargetRoot(
       `./dist-types/${packageDir}/src/index.d.ts`,
     );
   });
 
   let compilerState: CompilerState | undefined = undefined;
 
+  const allowWarningPkg = Array.isArray(allowWarnings) ? allowWarnings : [];
+
+  const messagesConf: { [key: string]: { logLevel: string } } = {};
+  for (const messageCode of omitMessages) {
+    messagesConf[messageCode] = {
+      logLevel: 'none',
+    };
+  }
   const warnings = new Array<string>();
 
   for (const packageDir of packageDirs) {
     console.log(`## Processing ${packageDir}`);
-    const projectFolder = resolvePath(process.cwd(), packageDir);
-    const packageFolder = resolvePath(
-      process.cwd(),
+    const noBail = Array.isArray(allowWarnings)
+      ? allowWarnings.some(aw => aw === packageDir || minimatch(packageDir, aw))
+      : allowWarnings;
+
+    const projectFolder = cliPaths.resolveTargetRoot(packageDir);
+    const packageFolder = cliPaths.resolveTargetRoot(
       './dist-types',
       packageDir,
     );
@@ -453,6 +400,7 @@ export async function runApiExtraction({
               logLevel: 'warning' as ExtractorLogLevel.Warning,
               addToApiReportFile: true,
             },
+            ...messagesConf,
           },
           tsdocMessageReporting: {
             default: {
@@ -543,12 +491,15 @@ export async function runApiExtraction({
     }
 
     const warningCountAfter = await countApiReportWarnings(projectFolder);
-    if (warningCountAfter > 0 && !ALLOW_WARNINGS.includes(packageDir)) {
+    if (noBail) {
+      console.log(`Skipping warnings check for ${packageDir}`);
+    }
+    if (warningCountAfter > 0 && !noBail) {
       throw new Error(
         `The API Report for ${packageDir} is not allowed to have warnings`,
       );
     }
-    if (warningCountAfter === 0 && ALLOW_WARNINGS.includes(packageDir)) {
+    if (warningCountAfter === 0 && allowWarningPkg.includes(packageDir)) {
       console.log(
         `No need to allow warnings for ${packageDir}, it does not have any`,
       );
@@ -1131,10 +1082,7 @@ export async function buildDocs({
   documenter.generateFiles();
 }
 
-export async function categorizePackageDirs(
-  projectRoot: string,
-  packageDirs: any[],
-) {
+export async function categorizePackageDirs(packageDirs: any[]) {
   const dirs = packageDirs.slice();
   const tsPackageDirs = new Array<string>();
   const cliPackageDirs = new Array<string>();
@@ -1150,7 +1098,7 @@ export async function categorizePackageDirs(
           }
 
           const pkgJson = await fs
-            .readJson(resolvePath(projectRoot, dir, 'package.json'))
+            .readJson(cliPaths.resolveTargetRoot(dir, 'package.json'))
             .catch(error => {
               if (error.code === 'ENOENT') {
                 return undefined;
@@ -1204,6 +1152,7 @@ function parseHelpPage(helpPageContent: string) {
 
   let options = new Array<string>();
   let commands = new Array<string>();
+  let commandArguments = new Array<string>();
 
   while (lines.length > 0) {
     while (lines.length > 0 && !lines[0].endsWith(':')) {
@@ -1228,6 +1177,8 @@ function parseHelpPage(helpPageContent: string) {
         options = sectionItems;
       } else if (sectionName?.toLocaleLowerCase('en-US') === 'commands:') {
         commands = sectionItems;
+      } else if (sectionName?.toLocaleLowerCase('en-US') === 'arguments:') {
+        commandArguments = sectionItems;
       } else {
         throw new Error(`Unknown CLI section: ${sectionName}`);
       }
@@ -1238,6 +1189,7 @@ function parseHelpPage(helpPageContent: string) {
     usage,
     options,
     commands,
+    commandArguments,
   };
 }
 
@@ -1249,6 +1201,7 @@ interface CliHelpPage {
   usage: string | undefined;
   options: string[];
   commands: string[];
+  commandArguments: string[];
 }
 
 async function exploreCliHelpPages(
@@ -1316,19 +1269,17 @@ function generateCliReport(name: string, models: CliModel[]): string {
 }
 
 interface CliExtractionOptions {
-  projectRoot: string;
   packageDirs: string[];
   isLocalBuild: boolean;
 }
 
 export async function runCliExtraction({
-  projectRoot,
   packageDirs,
   isLocalBuild,
 }: CliExtractionOptions) {
   for (const packageDir of packageDirs) {
     console.log(`## Processing ${packageDir}`);
-    const fullDir = resolvePath(projectRoot, packageDir);
+    const fullDir = cliPaths.resolveTargetRoot(packageDir);
     const pkgJson = await fs.readJson(resolvePath(fullDir, 'package.json'));
 
     if (!pkgJson.bin) {
@@ -1371,7 +1322,7 @@ export async function runCliExtraction({
           console.log('');
           console.log(
             `The conflicting file is ${relativePath(
-              projectRoot,
+              cliPaths.targetRoot,
               reportPath,
             )}, expecting the following content:`,
           );