+++ plugins/catalog-react/src/components/EntityOwnerPicker/EntityOwnerPicker.test.tsx
@@ -183,6 +183,7 @@ describe('<EntityOwnerPicker/>', () => {
         value={{
           updateFilters,
           queryParameters: { owners: ['team-a'] },
+          backendEntities: sampleEntities,
         }}
       >
         <EntityOwnerPicker />
@@ -196,6 +197,7 @@ describe('<EntityOwnerPicker/>', () => {
         value={{
           updateFilters,
           queryParameters: { owners: ['team-b'] },
+          backendEntities: sampleEntities,
         }}
       >
         <EntityOwnerPicker />
@@ -203,6 +205,23 @@ describe('<EntityOwnerPicker/>', () => {
     );
     expect(updateFilters).toHaveBeenLastCalledWith({
       owners: new EntityOwnerFilter(['team-b']),
+    });
+  });
+  it('removes owners from filters if there are none available', () => {
+    const updateFilters = jest.fn();
+    render(
+      <MockEntityListContextProvider
+        value={{
+          updateFilters,
+          queryParameters: { owners: ['team-a'] },
+          backendEntities: [],
+        }}
+      >
+        <EntityOwnerPicker />
+      </MockEntityListContextProvider>,
+    );
+    expect(updateFilters).toHaveBeenLastCalledWith({
+      owners: undefined,
     });
   });
 });