+++ plugins/catalog-customized/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @internal/plugin-catalog-customized
 
+## 0.0.5-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.7.0-next.3
+  - @backstage/plugin-catalog-react@1.2.2-next.3
+
 ## 0.0.5-next.2
 
 ### Patch Changes