+++ plugins/explore/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-explore
 
+## 0.3.43-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.3
+  - @backstage/catalog-model@1.1.4-next.1
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/errors@1.1.4-next.1
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.2.2-next.3
+  - @backstage/plugin-explore-common@0.0.1-next.0
+  - @backstage/plugin-explore-react@0.0.24-next.2
+  - @backstage/plugin-search-common@1.2.0-next.2
+  - @backstage/plugin-search-react@1.3.0-next.3
+
 ## 0.3.43-next.2
 
 ### Patch Changes