+++ plugins/firehydrant/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-firehydrant
 
+## 0.1.29-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.3
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.2.2-next.3
+
 ## 0.1.29-next.2
 
 ### Patch Changes