+++ plugins/github-deployments/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-github-deployments
 
+## 0.1.43-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.3
+  - @backstage/catalog-model@1.1.4-next.1
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/errors@1.1.4-next.1
+  - @backstage/integration@1.4.1-next.1
+  - @backstage/integration-react@1.1.7-next.3
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.2.2-next.3
+
 ## 0.1.43-next.2
 
 ### Patch Changes