+++ plugins/shortcuts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-shortcuts
 
+## 0.3.4-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.3
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/types@1.0.2-next.1
+
 ## 0.3.4-next.2
 
 ### Patch Changes