+++ plugins/stack-overflow/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-stack-overflow
 
+## 0.1.8-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.3
+  - @backstage/config@1.0.5-next.1
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-home@0.4.28-next.3
+  - @backstage/plugin-search-common@1.2.0-next.2
+
 ## 0.1.8-next.2
 
 ### Patch Changes